added img placeholder and buttons
