+++ src/App.js
@@ -4,7 +4,7 @@ import Developers from "./Components/Developers";
 import Footer from "./Components/Footer";
 import Header from "./Components/Header";
 import Home from "./Components/Home";
-//import Horoscope from './Components/Horoscope';
+import Horoscope from './Components/Horoscope';
 import "./App.css";
 
 function App() {
