+++ src/Components/Horoscope/index.css
@@ -1,0 +1,3 @@
+.btn{
+    width: 100%;
+}
