--- src/Components/Horoscope/index.js
@@ -1,29 +1,0 @@
-import React from 'react'
-
-function index() {
-    return (
-        <div>
-            <div className="sign-info">
-            </div>
-            <p>Sign Information</p>
-
-            <div className="yesterday">
-            <p>Yesterday</p>
-            </div>
-            <div className="today">
-            <p>Today</p>
-            </div>
-            <div className="tomorrow">
-            <p>Tomorrow</p>
-            </div>
-
-            <div className="daily-horoscope">
-                <p>horoscope</p>
-            </div>
-
-    
-        </div>
-    )
-}
-
-export default index
